Add Blowfish cipher implementation, wire up bcrypt to Blowfish/BLF API
- New file blowfish.c: OpenBSD-derived complete Blowfish block cipher
  implementation (Feistel network, S-boxes, P-Box,
  ECB/CBC/streaming/block APIs).
  All routines are marked DEF_WEAK for weak linking/override.
- New file blf.h: OpenBSD API for Blowfish context, initialization,
  key expansion, encryption/decryption routines,
  including Blowfish_encipher, Blowfish_expandstate,
  and convenience wrappers (blf_key, blf_enc, etc).
- bcrypt_int.c: Now includes blf.h and uses blf APIs
  for password hashing. Uncomments/activates all Blowfish logic
  required by bcrypt (replaces header stub).
- Fix bug in bcrypt_newhash: pass correct pointer type
  to bcrypt_initsalt.
- Now, bcrypt password hashes are securely computed using
  the complete Blowfish implementation, as in OpenBSD.

+++ portable/bcrypt_int.c
@@ -41,7 +41,7 @@
 #include <stdio.h>
 #include <stdint.h>
 #include <time.h>
-//#include "../blf_int.h"
+#include "../blf.h"
 #include "../pwd_int.h"
 #include "stdlib_int.h"
 #include "string_int.h"
@@ -211,7 +211,7 @@ bcrypt_newhash(const char *pass, int log_rounds, char *hash, size_t hashlen)
 {
 	char salt[BCRYPT_SALTSPACE];
 
-	if (bcrypt_initsalt(log_rounds, salt, sizeof(salt)) != 0)
+	if (bcrypt_initsalt(log_rounds, (uint8_t)salt, sizeof(salt)) != 0)
 		return -1;
 
 	if (bcrypt_hashpass(pass, salt, hash, hashlen) != 0)
